Add files via upload
What is a QR Code generator? It is a niche tool that is used to generate different types of QR Codes. Depending on your purpose, you can use our generator to create QR Codes to open a website, view a PDF file, listen to music, watch YouTube videos, store image files, connect to a WiFi network, and much more.